reorganize classes
