+++ src/main/java/de/eldoria/bigdoorsopener/conditions/DoorState.java
@@ -1,0 +1,3 @@
+package de.eldoria.bigdoorsopener.conditions;
+
+public enum DoorState {OPEN, CLOSED}
