+++ src/main/java/de/eldoria/bigdoorsopener/core/listener/registration/RegisterInteraction.java
@@ -1,4 +1,4 @@
-package de.eldoria.bigdoorsopener.listener.registration;
+package de.eldoria.bigdoorsopener.core.listener.registration;
 
 import de.eldoria.bigdoorsopener.config.Config;
 import de.eldoria.eldoutilities.messages.MessageSender;
