--- src/main/java/de/eldoria/bigdoorsopener/doors/conditions/DoorState.java
@@ -1,3 +1,0 @@
-package de.eldoria.bigdoorsopener.doors.conditions;
-
-public enum DoorState {OPEN, CLOSED}
